+++ tests/ApiTest.php
@@ -23,8 +23,8 @@ class ApiTest extends PHPUnit_Framework_TestCase
         $this->expectException(UnresolvedHandlerException::class);
 
         $api = $this->prepareApi('post');
-        $api->add('post', 'user/{id}', '#hander1');
-        $api->post('message/{id}', '#hander2');
+        $api->add('post', 'user/<id>', '#hander1');
+        $api->post('message/<id>', '#hander2');
 
         $api->run('user/10', $this->prepareLogger());
     }
@@ -36,54 +36,54 @@ class ApiTest extends PHPUnit_Framework_TestCase
     {
         $api = $this->prepareApi('get');
         $api->get('userGet', '#handler');
-        $api->get('userGet/{id}', '#handler');
+        $api->get('userGet/<id>', '#handler');
         $api->get('userGetPost', '#handler');
         $api->post('userPost', '#handler');
         $api->post('userGetPost', '#handler');
-        $api->post('userPost/{id}', '#handler');
+        $api->post('userPost/<id>', '#handler');
         $api->put('userPut', '#handler');
         $api->delete('userDelete', '#handler');
-        $api->put('userPut/{id}', '#handler');
-        $api->delete('userDelete/{id}', '#handler');
+        $api->put('userPut/<id>', '#handler');
+        $api->delete('userDelete/<id>', '#handler');
         $api->options('userOptions', '#handler');
         $api->patch('userPatch', '#handler');
-        $api->patch('userPatch/{id}', '#handler');
-        $api->options('userOptions/{id}', '#handler');
+        $api->patch('userPatch/<id>', '#handler');
+        $api->options('userOptions/<id>', '#handler');
         $api->add('purge', 'userPurge', '#handler');
-        $api->add('purge', 'userPurge/{id}', '#handler');
+        $api->add('purge', 'userPurge/<id>', '#handler');
         $api->add('get', 'userGetAdd', '#handler');
 
         $routeDefinition = [
             'get' => [
                 new Route('get', 'userGet', '#handler'),
-                new Route('get', 'userGet/{id}', '#handler'),
+                new Route('get', 'userGet/<id>', '#handler'),
                 new Route('get', 'userGetPost', '#handler'),
                 new Route('get', 'userGetAdd', '#handler')
             ],
             'post' => [
                 new Route('post', 'userPost', '#handler'),
                 new Route('post', 'userGetPost', '#handler'),
-                new Route('post', 'userPost/{id}', '#handler')
+                new Route('post', 'userPost/<id>', '#handler')
             ],
             'put' => [
                 new Route('put', 'userPut', '#handler'),
-                new Route('put', 'userPut/{id}', '#handler')
+                new Route('put', 'userPut/<id>', '#handler')
             ],
             'delete' => [
                 new Route('delete', 'userDelete', '#handler'),
-                new Route('delete', 'userDelete/{id}', '#handler')
+                new Route('delete', 'userDelete/<id>', '#handler')
             ],
             'options' => [
                 new Route('options', 'userOptions', '#handler'),
-                new Route('options', 'userOptions/{id}', '#handler')
+                new Route('options', 'userOptions/<id>', '#handler')
             ],
             'patch' => [
                 new Route('patch', 'userPatch', '#handler'),
-                new Route('patch', 'userPatch/{id}', '#handler')
+                new Route('patch', 'userPatch/<id>', '#handler')
             ],
             'purge' => [
                 new Route('purge', 'userPurge', '#handler'),
-                new Route('purge', 'userPurge/{id}', '#handler')
+                new Route('purge', 'userPurge/<id>', '#handler')
             ]
         ];